Added navigation bar
