+++ crwn-clothing/src/components/Directory/directory.component.jsx
@@ -1,4 +1,4 @@
-import CategoryItem from "../components/category-item/category-item.component";
+import CategoryItem from "../category-item/category-item.component";
 import "./directory.style.scss";
 
 const Directory = ({ categories }) => {
