+++ crwn-clothing/src/components/Directory/directory.style.scss
